backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
cryptography==41.0.4
python-dotenv==1.0.0
Flask-Cors==4.0.0
scikit-learn==1.4.2
numpy==1.26.4
